move Bundle to helperfuncs file; add another test and more documentation
