+++ pyrtl/__init__.py
@@ -18,7 +18,6 @@ from .wire import WireVector
 from .wire import Input, Output
 from .wire import Const
 from .wire import Register
-from .wire import Bundle
 
 # helper functions
 
@@ -38,6 +37,7 @@ from .helperfuncs import rtl_assert
 from .helperfuncs import check_rtl_assertions
 from .helperfuncs import find_loop
 from .helperfuncs import find_and_print_loop
+from .helperfuncs import Bundle
 
 from .corecircuits import and_all_bits
 from .corecircuits import or_all_bits
