+++ pyrtl/wire.py
@@ -8,7 +8,6 @@ Types defined in this file include:
 * `Output` -- a wire vector that defines an output for a block
 * `Const` -- a wire vector fed by a constant
 * `Register` -- a wire vector that is latched each cycle
-* `Bundle` -- a wire vector that has named fields for easy access
 """
 
 from __future__ import print_function, unicode_literals
@@ -17,7 +16,6 @@ import numbers
 import six
 import re
 import sys
-from functools import reduce
 
 from . import core  # needed for _setting_keep_wirevector_call_stack
 
@@ -512,6 +510,8 @@ class WireVector(object):
             return concat(extvector, self)
 
     def as_bundle(self, obj):
+        from .helperfuncs import Bundle
+
         bundle_bw = Bundle.get_bundle_bitwidth(obj)
 
         if len(self) != bundle_bw:
@@ -723,138 +723,3 @@ class Register(WireVector):
         self.reg_in = next
         net = LogicNet('r', None, args=(self.reg_in,), dests=(self,))
         working_block().add_net(net)
-
-
-class Bundle(WireVector):
-    """ A WireVector whose individual bits are named.
-
-    The initializer takes as its first argument the name of a class whose
-    attributes will be interpreted as the names and lengths of fields in a wire.
-    The order in which the attributes are defined is important; the first class
-    attribute is the MSB of the wire, and the last class attribute of the list is the LSB.
-
-    For example, say there is a wire that represents an instruction. If we wanted to name
-    certain segments of bits a certain way, we would create a class with the names and lengths
-    of these fields as attributes follows:
-
-        class RFormat:
-            funct7 = 7
-            rs2 = 5
-            rs1 = 5
-            funct3 = 3
-            rd = 5
-            opcode = 7
-
-    Then use it as the argument to Bundle to get back an object whose fields are actually
-    wirevectors, accessible by field name:
-
-        w = pyrtl.Bundle(RFormat)
-        w <<= 0b00000100110001010000010110010011
-        assert sim.inspect(w.funct7) == 0b0000010
-        assert sim.inspect(w.rs2) == 0b01100
-        assert sim.inspect(w.rs1) == 0b01010
-        assert sim.inspect(w.funct3) == 0b000
-        assert sim.inspect(w.rd) == 0b01011
-        assert sim.inspect(w.opcode) == 0b0010011
-
-    It can be used anywhere a normal wire can be used:
-
-        r = pyrtl.Register(len(w), "r")
-        r.next <<= w
-        # ...after stepping a few times...
-        assert sim.inspect(r) == 0b00000100110001010000010110010011
-
-    And you can interpret other wires as instances of the bundled class, by calling
-    `as_bundle`. This does lightweight checks such as making sure that the bundled class
-    and the wire you call `as_bundle` on has the same length so that the bits can map properly.
-    This allows you to access portions of the wire via fields.
-
-        f7 = r.as_bundle(RFormat).funct7
-        assert sim.inspect(f7) == 0b0000010
-
-        y = r.as_bundle(RFormat)
-        assert sim.inspect(y.funct7) == 0b0000010
-        assert sim.inspect(y.rs2) == 0b01100
-        assert sim.inspect(y.rs1) == 0b01010
-        assert sim.inspect(y.funct3) == 0b000
-        assert sim.inspect(y.rd) == 0b01011
-        assert sim.inspect(y.opcode) == 0b0010011
-
-    You can also pass in a list of (field, width) pairs:
-
-        rformat = [("funct7", 7), ("rs2", 5), ("rs1", 5), ("funct3", 3), ("rd", 5), ("opcode", 7)]
-        w = pyrtl.Bundle(rformat)
-
-    or an (ordered) dictionary (OrderedDict is the default for Python >= 3.7):
-
-        rformat = {"funct7": 7, "rs2": 5, "rs1": 5, "funct3": 3, "rd": 5, "opcode": 7}
-        w = pyrtl.Bundle(rformat)
-
-    instead of a class to form a Bundle. In all forms, order is important.
-
-    In all cases, the 'width' member may actually be a tuple of the form (n, w),
-    where n is the actual width and f is a wirevector or function returning
-    a wirevector that will be used to define the wire. Otherwise, 'width' should
-    just be an integer and will be interpreted as the literal width.
-    """
-    @staticmethod
-    def _get_fields(obj):
-        if isinstance(obj, list) and all(map(lambda t: isinstance(t, tuple), obj)):
-            # Passed in a list of tuples (i.e. (field, width) pairs), in order from MSB to LSB
-            fields = obj
-        elif isinstance(obj, dict):
-            from collections import OrderedDict
-            if (not (sys.version_info[0] >= 3 and sys.version_info[1] >= 7)
-               and (not isinstance(obj, OrderedDict))):
-                raise PyrtlError("For Python versions < 3.7, the dictionary used to instantiate "
-                                 "a Bundle must be explicitly ordered (i.e. OrderedDict)")
-            # Assume dictionary stores (field, width) pairs
-            fields = list(obj.items())
-        elif isinstance(obj, six.class_types):
-            if not (sys.version_info[0] >= 3 and sys.version_info[1] >= 7):
-                raise PyrtlError("Passing a class as an argument to Bundle() "
-                                 "is only allowed for Python versions >= 3.7")
-            # Let's assume 'obj' is a **class** name, so treat it as if it has field names.
-            # As of Python 3.7, dictionaries preserve insertion order, so a class's attributes
-            # (in __dict__) will being ordered as well. This relies on that fact because the
-            # fields are defined in MSB to LSB order in the class.
-            fs = filter(lambda attr: not attr.startswith("__"), vars(obj))
-            fields = [(attr, getattr(obj, attr)) for attr in fs]
-        else:
-            raise PyrtlError("Cannot determine (field, width) pairs from %s object" % type(obj))
-        return fields
-
-    @staticmethod
-    def get_bundle_bitwidth(obj):
-        fields = Bundle._get_fields(obj)
-
-        def aux(acc, t):
-            if isinstance(t[1], tuple):
-                width = t[1][0]
-            else:
-                width = t[1]
-            return acc + width
-        return reduce(aux, fields, 0)
-
-    def __init__(self, obj, name="", block=None):
-        super(Bundle, self).__init__(Bundle.get_bundle_bitwidth(obj), name, block)
-
-        fields = Bundle._get_fields(obj)
-        start = 0
-        args = []
-        for field, length in fields[::-1]:
-            if isinstance(length, tuple):
-                from .corecircuits import as_wires
-                # length is actually a tuple of the form (width, val)
-                val = length[1]
-                length = length[0]
-                if callable(val):
-                    val = val()
-                val = as_wires(val, bitwidth=length)
-                args.append(val)
-            setattr(self, field, self[start:start + length])
-            start += length
-
-        if args:
-            from .corecircuits import concat_list
-            self <<= concat_list(args)
